# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Evolution GO - whatsmeow",
        "title": "Evolution GO",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/send/contact": {
            "post": {
                "description": "Send a contact message",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Send Message"
                ],
                "summary": "Send a contact message",
                "parameters": [
                    {
                        "description": "Message data",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pkg_sendMessage_handler.SendContactMessageRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "400": {
                        "description": "Error on validation",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/send/link": {
            "post": {
                "description": "Send a link message",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Send Message"
                ],
                "summary": "Send a link message",
                "parameters": [
                    {
                        "description": "Message data",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pkg_sendMessage_handler.SendTextMessageRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "400": {
                        "description": "Error on validation",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/send/list": {
            "post": {
                "description": "Send a list message",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Send Message"
                ],
                "summary": "Send a list message",
                "parameters": [
                    {
                        "description": "Message data",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pkg_sendMessage_handler.SendListMessageRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "400": {
                        "description": "Error on validation",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/send/location": {
            "post": {
                "description": "Send a location message",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Send Message"
                ],
                "summary": "Send a location message",
                "parameters": [
                    {
                        "description": "Message data",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pkg_sendMessage_handler.SendLocationMessageRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "400": {
                        "description": "Error on validation",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/send/media": {
            "post": {
                "description": "Send a media message",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Send Message"
                ],
                "summary": "Send a media message",
                "parameters": [
                    {
                        "description": "Message data",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pkg_sendMessage_handler.SendMediaMessageRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "400": {
                        "description": "Error on validation",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/send/poll": {
            "post": {
                "description": "Send a poll message",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Send Message"
                ],
                "summary": "Send a poll message",
                "parameters": [
                    {
                        "description": "Message data",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pkg_sendMessage_handler.SendPollMessageRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "400": {
                        "description": "Error on validation",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/send/sticker": {
            "post": {
                "description": "Send a sticker message",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Send Message"
                ],
                "summary": "Send a sticker message",
                "parameters": [
                    {
                        "description": "Message data",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pkg_sendMessage_handler.SendStickerMessageRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "400": {
                        "description": "Error on validation",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/send/text": {
            "post": {
                "description": "Send a text message",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Send Message"
                ],
                "summary": "Send a text message",
                "parameters": [
                    {
                        "description": "Message data",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pkg_sendMessage_handler.SendTextMessageRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "400": {
                        "description": "Error on validation",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "gin.H": {
            "type": "object",
            "additionalProperties": {}
        },
...
        "pkg_sendMessage_handler.SendContactMessageRequest": {
            "type": "object",
            "properties": {
                "phone": {
                    "type": "string",
                    "example": "557499879409"
                },
                "vcard": {
                    "type": "object",
                    "properties": {
                        "fullName": {
                            "type": "string",
                            "example": "John Doe"
                        },
                        "phone": {
                            "type": "string",
                            "example": "557499879409"
                        }
                    }
                }
            }
        },
        "pkg_sendMessage_handler.SendListMessageRequest": {
            "type": "object",
            "properties": {
                "buttonText": {
                    "type": "string",
                    "example": "View"
                },
                "desc": {
                    "type": "string",
                    "example": "This is a list"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "[\"Item 1\"",
                        " \"Item 2\"]"
                    ]
                },
                "phone": {
                    "type": "string",
                    "example": "557499879409"
                }
            }
        },
        "pkg_sendMessage_handler.SendLocationMessageRequest": {
            "type": "object",
            "properties": {
                "latitude": {
                    "type": "number",
                    "example": -23.5505199
                },
                "longitude": {
                    "type": "number",
                    "example": -46.6333094
                },
                "name": {
                    "type": "string",
                    "example": "São Paulo"
                },
                "phone": {
                    "type": "string",
                    "example": "557499879409"
                }
            }
        },
        "pkg_sendMessage_handler.SendMediaMessageRequest": {
            "type": "object",
            "properties": {
                "phone": {
                    "type": "string",
                    "example": "557499879409"
                },
                "type": {
                    "type": "string",
                    "example": "image"
                },
                "url": {
                    "type": "string",
                    "example": "https://example.com/image.jpg"
                }
            }
        },
        "pkg_sendMessage_handler.SendPollMessageRequest": {
            "type": "object",
            "properties": {
                "maxAnswer": {
                    "type": "integer",
                    "example": 1
                },
                "options": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "[\"Red\"",
                        " \"Blue\"]"
                    ]
                },
                "phone": {
                    "type": "string",
                    "example": "557499879409"
                },
                "question": {
                    "type": "string",
                    "example": "What is your favorite color?"
                }
            }
        },
        "pkg_sendMessage_handler.SendStickerMessageRequest": {
            "type": "object",
            "properties": {
                "phone": {
                    "type": "string",
                    "example": "557499879409"
                },
                "sticker": {
                    "type": "string",
                    "example": "https://example.com/sticker.webp"
                }
            }
        },
        "pkg_sendMessage_handler.SendTextMessageRequest": {
            "type": "object",
            "properties": {
                "phone": {
                    "type": "string",
                    "example": "557499879409"
                },
                "text": {
                    "type": "string",
                    "example": "Teste"
                }
            }
        },
...
    }
}